Move comprobar_sucesion to Algoritmo
